feat: check the input len is a power of 2
